#25 Add Skeleton Loader + Image load
I have added a skeleton loader and an image fallback, as the backend sometimes fails to load at all. Need to reset the local host to make it work again. Skeleton lay-out is the basic lay-out of the page and I have tested this with a javascript timer. (Deleted this, but does work when loading takes time)